requirements.txt
aioredis==2.0.1          # Latest stable version
asgiref==3.7.2           # Compatible with Django 4.x
attrs==23.1.0            # General-purpose library
autobahn==23.1.2         # WebSocket support
certifi==2023.11.17      # SSL Certificates
cffi==1.15.1             # Cryptography library
channels==4.0.0          # Updated for Django 4.x
channels-redis==4.0.0    # Compatible with latest channels
cryptography==41.0.3     # Security updates
daphne==4.0.0            # ASGI server
defusedxml==0.7.1        # XML security
Django==4.2.8            # Long-term support version
django-allauth==0.54.0   # Updated social authentication
django-rest-auth==0.9.5  # Updated REST authentication
djangorestframework==3.14.0  # API framework
drf-extensions==0.7.0    # DRF extensions
gunicorn==20.1.0         # WSGI HTTP server
hiredis==2.0.0           # Redis support
oauthlib==3.2.2          # OAuth 2.0 authentication
psycopg2-binary==2.9.8   # PostgreSQL driver
PyJWT==2.8.0             # JWT tokens
pytz==2023.3             # Timezone utilities
redis==5.0.0             # Redis library
requests==2.31.0         # HTTP requests
requests-oauthlib==1.3.1 # OAuth integration
six==1.16.0              # Python 2/3 compatibility
tqdm==4.66.1             # Progress bar
websockets==11.0.3       # WebSocket support
